extract import map generation logic into utility (#668)
* extract package walk logic into utility

* lower thresholds

* remove removed config

* create import map merging function and apply to graphql plugin

* 538 refactor rollup and resource to use shared data structure

* post rebase fixes
